backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
rembg
onnxruntime
onnxruntime-gpu
